* commit 'c2f68f624db36708a954424f2a154ea3c91d0f4b':
  Update the readme
  Update procfile command
  Fixes for buckets as an NPM module
  Change minify task to prepublish for Heroku
  Nix prepublish for now
  Move packages back to dependencies
  Refactor express app for NPM use & Express 4 (WIP)

+++ test/reset.coffee
@@ -1,17 +1,25 @@
-mongoose = require 'mongoose'
-
 db = require '../server/lib/database'
 config = require '../server/config'
+buckets = require('../server')
 Entry = require '../server/models/entry'
 
 dropDatabase = (done) ->
-  db.connection.db.dropDatabase done
+  prep ->
+    db.connection.db.dropDatabase done
 
 prep = (done) ->
-  mongoose.connection.on 'connected', done
+  if db.connection.readyState is 1
+    done()
+  else
+    db.connection.on 'connected', ->
+      db.connection.db.dropDatabase done
+
+server = (done) ->
+  buckets().start done
 
 module.exports =
   db: dropDatabase
   prep: prep
+  server: server
   all: (done) ->
     dropDatabase done
